+++ generators/app/index.js
@@ -148,15 +148,38 @@ module.exports = class extends Generator {
       },
       {
         type: 'list',
+        name: 'overleaf',
+        message: 'Overleaf compatibility?',
+        choices: [
+          {
+            name: "yes",
+            value: true
+          },
+          {
+            name: "no",
+            value: false
+          }
+        ],
+        default: true
+      },
+      {
+        type: 'list',
         name: 'texlive',
-        message: 'Which TeXLive compatiblity?',
+        when: function(response) {
+          return (!response.overleaf || (response.overleaf == "false"));
+        },
+        message: 'Which TeXLive compatibility?',
         choices: [
           {
             name: "TeXLive 2021",
             value: 2021
+          },
+          {
+            name: "TeXLive 2022",
+            value: 2022
           }
         ],
-        default: 2021
+        default: 2022
       },
       {
         type: 'list',
@@ -165,7 +188,7 @@ module.exports = class extends Generator {
         choices: ["pdflatex", "lualatex"],
         default: "pdflatex",
         when: function(response) {
-          return !((response.documentclass === 'ieee') && (response.texlive == 2021));
+          return !(response.documentclass === 'ieee');
         }
       },
       {
@@ -394,6 +417,11 @@ module.exports = class extends Generator {
       // To access props later use this.props.someAnswer;
       this.props = props;
 
+      if (this.props.overleaf) {
+        // we do not prompt for texlive version in case of overleaf
+        this.props.texlive = 2021;
+      }
+
       // somehow texlive is not routed through
       // special handling
       if (this.params.options.texlive) {
@@ -428,6 +456,7 @@ module.exports = class extends Generator {
       this.props.cleveref = (this.props.cleveref === true) || (this.props.cleveref === 'true')
       this.props.examples = (this.props.examples === true) || (this.props.examples === 'true')
       this.props.howtotext = (this.props.howtotext === true) || (this.props.howtotext === 'true')
+      this.props.overleaf = (this.props.overleaf === true) || (this.props.overleaf === 'true')
 
       if (this.props.examples) {
         this.props.useExampleEnvironment = true;
@@ -513,7 +542,7 @@ module.exports = class extends Generator {
     );
     global.fs.copyTpl(
       global.templatePath('latexmkrc'),
-      global.destinationPath('latexmkrc'),
+      global.destinationPath(global.props.overleaf ? '_latexmkrc' : 'latexmkrc'),
       global.props
     );
     global.fs.copyTpl(
@@ -533,8 +562,8 @@ module.exports = class extends Generator {
     );
     if (global.props.documentclass === 'lncs') {
       global.fs.copy(
-        global.templatePath('splncsnat.bst'),
-        global.destinationPath('splncsnat.bst')
+        global.templatePath('splncs04nat.bst'),
+        global.destinationPath('splncs04nat.bst')
       );
     }
     if (global.props.language === 'de') {